Split analysis/mod.rs into focused orchestration sub-modules (#562)
Decompose the 1,830-line analysis/mod.rs into three focused sub-modules:
- orchestration.rs: analyze_all() entry point and deadline ordering
- candidate_aggregation.rs: candidate merging and post-processing
- module_dispatch_specs.rs: discovery module spec builders and clustering

The original mod.rs is reduced to 266 lines of module declarations and
re-exports. All public API remains unchanged for FFI consumers.
